MDS-448 | More test updates
