+++ test/component/uk/gov/hmrc/individualsdetailsapi/controllers/LiveAddressesControllerSpec.scala
@@ -68,34 +68,4 @@ class LiveAddressesControllerSpec extends CommonControllerWithIfRequestSpec {
      |  } ]
      |}""".stripMargin)
 
-  override val expectedJsonInvalidIf: JsValue = Json.parse(s"""{
-      |  "_links" : {
-      |    "self" : {
-      |      "href" : "/individuals/details/addresses?matchId=$matchId"
-      |    }
-      |  },
-      |  "residences" : [ ]
-      |}""".stripMargin)
-
-  scenario(s"user does not have valid scopes") {
-    Given("A valid auth token but invalid scopes")
-    AuthStub.willNotAuthorizePrivilegedAuthTokenNoScopes(authToken)
-
-    And("a valid record in the matching API")
-    IndividualsMatchingApiStub.hasMatchFor(matchId.toString, nino)
-
-    And("IF will return response")
-    IfStub.searchDetails(nino, ifDetailsResponse)
-
-    When(
-      s"I make a call to ${if (endpoint.isEmpty) "root" else endpoint} endpoint")
-    val response = invokeEndpoint(s"$serviceUrl/${endpoint}?matchId=$matchId")
-
-    Then("The response status should be 401")
-    response.code shouldBe UNAUTHORIZED
-    Json.parse(response.body) shouldBe Json.obj(
-      "code" -> "UNAUTHORIZED",
-      "message" ->"User does not have valid scopes"
-    )
-  }
 }
