+++ test/component/uk/gov/hmrc/individualsdetailsapi/stubs/IfStub.scala
@@ -17,14 +17,13 @@
 package component.uk.gov.hmrc.individualsdetailsapi.stubs
 
 import com.github.tomakehurst.wiremock.client.WireMock._
-import play.api.libs.json.Json
+import play.api.libs.json.{JsValue, Json}
 import play.api.test.Helpers._
-import uk.gov.hmrc.individualsdetailsapi.domain.integrationframework.{
-  IfDetails,
-  IfDetailsResponse
-}
+import uk.gov.hmrc.individualsdetailsapi.domain.integrationframework.{IfDetails, IfDetailsResponse}
 
 object IfStub extends MockHost(22004) {
+
+
 
   def searchDetails(nino: String, ifDetailsResponse: IfDetailsResponse) =
     mock.register(
@@ -33,6 +32,14 @@ object IfStub extends MockHost(22004) {
           aResponse()
             .withStatus(OK)
             .withBody(Json.toJson(ifDetailsResponse).toString())))
+
+  def customResponse(nino: String, status: Int, response: JsValue) =
+    mock.register(
+      get(urlPathEqualTo(s"/individuals/details/contact/nino/$nino"))
+        .willReturn(
+          aResponse()
+            .withStatus(status)
+            .withBody(response.toString())))
 
   def enforceRateLimit(nino: String, fromDate: String, toDate: String): Unit =
     mock.register(
